Create Day-8(P-4)
Given an odd-digit-number N, you need to find whether the given number is a balanced number or not. An odd-digit number is called a balanced number if the sum of all digits to the left of the middle digit and the sum of all digits to the right of the middle digit is equal. Note:- The number of digits of 1234006. is 7, i.e. odd digits. The middle digit is 4. LHS sum is 1+2+3=6, and RHS sum is 0+0+6=6.